Extract analysis type methods to a module and refactor Post#raw_links
